[2.0] Zend_Ldap tests now pass
git-svn-id: http://framework.zend.com/svn/framework/standard/branches/development-2.0@21434 44c647ce-9c0f-0410-b52a-842ac1e357ba

+++ tests/Zend/Ldap/CopyRenameTest.php
@@ -101,6 +101,9 @@ class Zend_Ldap_CopyRenameTest extends Zend_Ldap_OnlineTestCase
 
     protected function tearDown()
     {
+        if (!constant('TESTS_ZEND_LDAP_ONLINE_ENABLED')) {
+            return;
+        }
         if ($this->_getLdap()->exists($this->_newDn))
             $this->_getLdap()->delete($this->_newDn, false);
         if ($this->_getLdap()->exists($this->_orgDn))
